Move locks into separate file
