+++ packages/sdk/client-services/src/packlets/locks/index.ts
@@ -1,0 +1,5 @@
+//
+// Copyright 2023 DXOS.org
+//
+
+export const createLock = ()
