+++ packages/sdk/client-services/src/packlets/locks/node-resource-lock.ts
