+++ packages/sdk/client-services/src/packlets/locks/resource-lock.ts
