+++ packages/sdk/client-services/src/packlets/services/index.ts
@@ -6,5 +6,3 @@ export * from './local-client-services';
 export * from './service-context';
 export * from './service-host';
 export * from './service-registry';
-export * from './resource-lock';
-export * from './node-resource-lock';
