+++ packages/sdk/client-services/src/packlets/services/service-host.ts
@@ -28,8 +28,8 @@ import { SpacesServiceImpl } from '../spaces';
 import { createStorageObjects } from '../storage';
 import { SystemServiceImpl } from '../system';
 import { VaultResourceLock } from '../vault';
-import { NodeResourceLock } from './node-resource-lock';
-import { ResourceLock, ResourceLockOptions } from './resource-lock';
+import { NodeResourceLock } from '../locks/node-resource-lock';
+import { ResourceLock, ResourceLockOptions } from '../locks/resource-lock';
 import { ServiceContext } from './service-context';
 import { ServiceRegistry } from './service-registry';
