+++ packages/sdk/client-services/src/packlets/vault/index.ts
@@ -5,6 +5,5 @@
 export * from './iframe-host-runtime';
 export * from './iframe-proxy-runtime';
 export * from './shell-runtime';
-export * from './vault-resource-lock';
 export * from './worker-runtime';
 export * from './worker-session';
